refactor: move request list components into subfolder
