+++ src/views/Generator/GeneratorTabs/RequestList/Header.tsx
@@ -4,7 +4,7 @@ import { useGeneratorStore } from '@/store/generator'
 import { getFileNameWithoutExtension } from '@/utils/file'
 import { RecorderIcon } from '@/components/icons'
 
-export function RequestListHeader({
+export function Header({
   filter,
   setFilter,
   filterAllData,
