+++ src/views/Generator/GeneratorTabs/RequestList/RequestList.utils.tsx
@@ -3,7 +3,7 @@ import { ProxyData } from '@/types'
 import { GlobeIcon } from '@radix-ui/react-icons'
 import { Button } from '@radix-ui/themes'
 import { ComponentProps } from 'react'
-import { RecordingSelector } from '../RecordingSelector'
+import { RecordingSelector } from '../../RecordingSelector'
 
 export function validateRecording({
   allowlist,
