+++ src/views/Generator/GeneratorTabs/RequestList/index.ts
@@ -1,0 +1,1 @@
+export * from './RequestList'
